fix deleted task
