+++ src/App.tsx
@@ -1,10 +1,11 @@
-import React, { useState, useContext} from "react";
+import React, { useState, useContext } from "react";
 import styles from "./App.module.css";
 import TaskList from "./components/TaskList/TaskList";
 import { TaskProvider } from "./context/TaskContext";
 import AddTaskForm from "./components/AddTaskForm/AddTaskForm";
 import { ThemeContext } from "./context/ThemeContext";
-import ThemeToggleButton from "./components/ThemeToggleButton/ThemeToggleButton"
+import ThemeToggleButton from "./components/ThemeToggleButton/ThemeToggleButton";
+import UndoButton from "./components/UndoButton/UndoButton";
 
 const App: React.FC = () => {
   const [filter, setFilter] = useState<"All" | "Completed" | "Incomplete">("All");
@@ -31,7 +32,7 @@ const App: React.FC = () => {
           />
           <div className={styles.filterContainer}>
             <button onClick={() => setIsFilterOpen(!isFilterOpen)} className={styles.filterButton}>
-            {filter.toUpperCase()}
+              {filter.toUpperCase()}
             </button>
             {isFilterOpen && (
               <div className={styles.filterDropdown}>
@@ -41,10 +42,15 @@ const App: React.FC = () => {
               </div>
             )}
           </div>
-          <ThemeToggleButton/>
+          <ThemeToggleButton />
         </div>
         <TaskList filter={filter} searchQuery={searchQuery} />
-        <AddTaskForm />
+        <div className={styles.buttonsContainer}>
+          <div>
+            <UndoButton />
+          </div>
+          <AddTaskForm />
+        </div>
       </div>
     </TaskProvider>
   );
