+++ src/components/AddTaskForm/AddTaskForm.module.css
@@ -1,7 +1,4 @@
 .addTaskButton {
-  position: absolute;
-  margin-top: 50px;
-  right: 25%;
   font-size: 36px;
   background-color: #218838;
   color: white;
