+++ src/components/TaskItem/TaskItem.module.css
@@ -92,25 +92,7 @@
   height: 100%;
   border-radius: 50%;
   background-color: #dc3545;
-  transition: width 1s linear, height 1s linear;
-}
-
-.undoButton {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  background-color: #ffcc00;
-  color: #000; 
-  border: none; 
-  padding: 10px 20px;
-  border-radius: 4px; 
-  cursor: pointer; 
-  font-size: 16px; 
-  transition: background-color 0.3s ease; 
-  transform: translate(-50%, -50%); 
-  z-index: 1000;
-}
-
-.undoButton:hover {
-  background-color: #e6b800;
+  transition:
+    width 1s linear,
+    height 1s linear;
 }
