+++ src/components/UndoButton/UndoButton.module.css
@@ -1,0 +1,14 @@
+.undoButton {
+  background-color: #ffcc00;
+  color: #000;
+  border: none;
+  padding: 10px 20px;
+  border-radius: 4px;
+  cursor: pointer;
+  font-size: 16px;
+  transition: background-color 0.3s ease;
+}
+
+.undoButton:hover {
+  background-color: #e6b800;
+}
